[edit]240130_Character
- 캐릭터 체력 int -> float

+++ Assets/Scripts/Player/PlayerStat.cs
@@ -14,7 +14,7 @@ public enum StatsChangeType
 public class PlayerStat
 {
     public StatsChangeType statsChangeType;
-    [Range(1, 100)] public int maxHealth;
+    [Range(1, 100)] public float maxHealth;
     [Range(1f, 20f)] public float speed;
     public AttackSO attackSO;
 }
